+++ src/lib.rs
@@ -63,6 +63,52 @@ impl Rtc {
         // of a PL031 device which is appropriately mapped.
         unsafe { addr_of_mut!((*self.registers).lr).write_volatile(unix_time) }
     }
+
+    /// Writes a match value. When the RTC value matches this then an interrupt
+    /// will be generated (if it is enabled).
+    pub fn set_match_timestamp(&mut self, match_timestamp: u32) {
+        // SAFETY: We know that self.registers points to the control registers
+        // of a PL031 device which is appropriately mapped.
+        unsafe { addr_of_mut!((*self.registers).mr).write_volatile(match_timestamp) }
+    }
+
+    /// Returns whether the match register matches the RTC value, whether or not
+    /// the interrupt is enabled.
+    pub fn matched(&self) -> bool {
+        // SAFETY: We know that self.registers points to the control registers
+        // of a PL031 device which is appropriately mapped.
+        let ris = unsafe { addr_of!((*self.registers).ris).read_volatile() };
+        (ris & 0x01) != 0
+    }
+
+    /// Returns whether there is currently an interrupt pending.
+    ///
+    /// This should be true if and only if `matched` returns true and the
+    /// interrupt is masked.
+    pub fn interrupt_pending(&self) -> bool {
+        // SAFETY: We know that self.registers points to the control registers
+        // of a PL031 device which is appropriately mapped.
+        let ris = unsafe { addr_of!((*self.registers).mis).read_volatile() };
+        (ris & 0x01) != 0
+    }
+
+    /// Sets or clears the interrupt mask.
+    ///
+    /// When the mask is true the interrupt is enabled; when it is false the
+    /// interrupt is disabled.
+    pub fn enable_interrupt(&mut self, mask: bool) {
+        let imsc = if mask { 0x01 } else { 0x00 };
+        // SAFETY: We know that self.registers points to the control registers
+        // of a PL031 device which is appropriately mapped.
+        unsafe { addr_of_mut!((*self.registers).imsc).write_volatile(imsc) }
+    }
+
+    /// Clears a pending interrupt, if any.
+    pub fn clear_interrupt(&mut self) {
+        // SAFETY: We know that self.registers points to the control registers
+        // of a PL031 device which is appropriately mapped.
+        unsafe { addr_of_mut!((*self.registers).icr).write_volatile(0x01) }
+    }
 }
 
 // SAFETY: `Rtc` just contains a pointer to device memory, which can be accessed from any context.